requirements.txt
dostoevsky==0.6.0
Flask==2.0.3
Flask-MySQL==1.5.2
googletrans==3.1.0a0
requests==2.27.1
fasttext==0.9.2